refactor maas request external_request class into module
